LPS-74160 Create base model listener class so that logic can be reused for user export
