+++ modules/apps/foundation/portal-security/portal-security-ldap/src/main/java/com/liferay/portal/security/ldap/internal/model/listener/BaseLDAPExportListener.java
@@ -1,0 +1,80 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.security.ldap.internal.model.listener;
+
+import com.liferay.portal.kernel.model.BaseModel;
+import com.liferay.portal.kernel.model.BaseModelListener;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.ldap.LDAPSettings;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.transaction.TransactionCommitCallbackUtil;
+import com.liferay.portal.security.exportimport.UserExporter;
+import com.liferay.portal.security.ldap.internal.UserImportTransactionThreadLocal;
+
+import java.io.Serializable;
+
+import java.util.Map;
+import java.util.concurrent.Callable;
+
+/**
+ * @author Minhchau Dang
+ */
+public abstract class BaseLDAPExportListener<T extends BaseModel<T>>
+	extends BaseModelListener<T> {
+
+	protected void exportToLDAP(
+			final User user, final UserExporter userExporter,
+			final LDAPSettings ldapSettings)
+		throws Exception {
+
+		if (user.isDefaultUser() ||
+			UserImportTransactionThreadLocal.isOriginatesFromImport()) {
+
+			return;
+		}
+
+		Callable<Void> callable = new Callable<Void>() {
+
+			@Override
+			public Void call() throws Exception {
+				ServiceContext serviceContext =
+					ServiceContextThreadLocal.getServiceContext();
+
+				Map<String, Serializable> expandoBridgeAttributes = null;
+
+				if (serviceContext != null) {
+					expandoBridgeAttributes =
+						serviceContext.getExpandoBridgeAttributes();
+				}
+
+				userExporter.exportUser(user, expandoBridgeAttributes);
+
+				return null;
+			}
+
+		};
+
+		if (ldapSettings.isPasswordPolicyEnabled(user.getCompanyId()) &&
+			user.isPasswordModified()) {
+
+			callable.call();
+		}
+		else {
+			TransactionCommitCallbackUtil.registerCallback(callable);
+		}
+	}
+
+}
