+++ modules/apps/foundation/portal-security/portal-security-ldap/src/main/java/com/liferay/portal/security/ldap/internal/model/listener/UserModelListener.java
@@ -18,7 +18,6 @@ import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.MembershipRequest;
 import com.liferay.portal.kernel.model.MembershipRequestConstants;
@@ -28,18 +27,12 @@ import com.liferay.portal.kernel.security.auth.PrincipalThreadLocal;
 import com.liferay.portal.kernel.security.ldap.LDAPSettings;
 import com.liferay.portal.kernel.service.MembershipRequestLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.portal.kernel.transaction.TransactionCommitCallbackUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.security.exportimport.UserExporter;
 import com.liferay.portal.security.ldap.internal.UserImportTransactionThreadLocal;
 
-import java.io.Serializable;
-
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.Callable;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -52,7 +45,7 @@ import org.osgi.service.component.annotations.ReferencePolicyOption;
  * @author Vilmos Papp
  */
 @Component(immediate = true, service = ModelListener.class)
-public class UserModelListener extends BaseModelListener<User> {
+public class UserModelListener extends BaseLDAPExportListener<User> {
 
 	@Override
 	public void onAfterAddAssociation(
@@ -106,41 +99,7 @@ public class UserModelListener extends BaseModelListener<User> {
 	}
 
 	protected void exportToLDAP(final User user) throws Exception {
-		if (user.isDefaultUser() ||
-			UserImportTransactionThreadLocal.isOriginatesFromImport()) {
-
-			return;
-		}
-
-		Callable<Void> callable = new Callable<Void>() {
-
-			@Override
-			public Void call() throws Exception {
-				ServiceContext serviceContext =
-					ServiceContextThreadLocal.getServiceContext();
-
-				Map<String, Serializable> expandoBridgeAttributes = null;
-
-				if (serviceContext != null) {
-					expandoBridgeAttributes =
-						serviceContext.getExpandoBridgeAttributes();
-				}
-
-				_userExporter.exportUser(user, expandoBridgeAttributes);
-
-				return null;
-			}
-
-		};
-
-		if (_ldapSettings.isPasswordPolicyEnabled(user.getCompanyId()) &&
-			user.isPasswordModified()) {
-
-			callable.call();
-		}
-		else {
-			TransactionCommitCallbackUtil.registerCallback(callable);
-		}
+		exportToLDAP(user, _userExporter, _ldapSettings);
 	}
 
 	protected void updateMembershipRequestStatus(long userId, long groupId)
